SOC-1454: People can Accept/Refuse a connection request between 2 others
What is the problem to fix?
Invitations to connect are displayed in the activity stream. 
A link to accept/refuse the invitation is displayed. 
Anybody viewing the stream can click on it to accept/refuse the invitation.

Fix description
The accept and refuse links are displayed only if the current user is the receiver of the connection request.