SOC-1519
What is the problem to fix?

    * Follow http://int.exoplatform.org/portal/public/intranet/forum/topic/topica83aff6e2e8902a900706f4795eff45e

When an user accepts a connection, the activity message displays the login instead of the name of the user :
thomas_delhomenie is now connected with patrice_lamarque.

Should be
Thomas Delhoménie is now connected with Patrice Lamarque.

How is the problem fixed?

    * The problem: from UIRelationshipActivity:

        private String getProfileLink(String username) {
          LinkProvider linkProvider = getApplicationComponent(LinkProvider.class);
          return "<a href=" + linkProvider.getProfileUri(username) + ">" + username +"</a>";
        }

To fix it, use LinkProvider#getProfileLink(String) instead:

  private String getProfileLink(String username) {
    LinkProvider linkProvider = getApplicationComponent(LinkProvider.class);
    return linkProvider.getProfileLink(username);
  }