SOC-5813 : define explicitly Elasticsearch mapping for space and profile types (#272)
Mappings definitions have been changed to raw strings instead of using JSONObject, for a better readability.
Date fields have been mapped as date.
Fields used for sorting in queries have been mapped 2 times : one with the 'text' type for the search, and one with the 'keyword' type for sorting. For example, for the profile, 'lastName' is the analyzed text field, and 'lastName.raw' is the raw value of the field indexed as keyword.