SOC-5775 : use new analyzer with whitespace tokenizer for profiles indexing
The tokenizer used until now was the standard tokenizer, which breaks words on whitespaces and some special characters like '-'.
It is a problem for profile names with such characters, like "Anne-Marie" for example. In such a case, it is indexed as "anne" and "marie" and cnnot be found when searching with "anne-marie".
The fix uses a new tokenizer (whitespace) which breaks words only on whitespaces.
It is used besides the standard tokenizer in profile fields, which means fields will be indexed in both ways. So "Anne-Marie" will be indexed as "anne", "marie" and "anne-marie".
Then searches in Profile page and Mentioning will use this new indexed version to allow to find "Anne-Marie" with the search text "Anne-Marie" or "Anne-Mar".