PLF-8026 Make sure the CREATED_DATE column of the table SOC_SPACES has a non zero default value (#381)
Since MySQL 5.7.7, the default strict mode has changed and does not allow zero value ('0000-00-00 00:00:00') for timestamps anymore.
Since this column definition has changed over time, we can be in a situation where we tried to migrate from a version having zero as the default value. in this case an error is raised during any table update (here for the addition of a new column BANNER_LAST_UPDATED).
This fix makes sure the default value for the column CREATED_DATE is a non zero value (now).