SOC-6074 : initialize switch button in Space members administration correctly (#403)
The switch buttons used in Spaces members administration are built with the js lib iOSCheckbox. This lib is defined in Gatein and has been duplicated in Social.
When initializing the button during the first display of the space members page, the lib instances from Gatein is executed (instead of the one from Social). Since this instance is configured differently (the CSS classes differs), the buttons were not initialized correctly. After a refresh, the Social instance of the lib was used and therefore the buttons were correctly renders.

Since the difference between the 2 lib instances is only about configuration, this fix removes the Social version of the lib, uses the Gatein version and initialize it with the right CSS classes for the Spaces members page in order to initialize correctly the buttons.