SOC-6139 : securize REST endpoints to suggest spaces and get spaces memberships (#437)
The 2 REST endpoints which allows to suggest spaces and get spaces memberships did not check if the current user is allowed to perform these operations.
Therefore it was possible to get the list of spaces, even without being authenticated.

This fix adds the needed authorization checks on these 2 endpoints to make sure the authenticated user tries to fetch only his/her own data, and that the authenticated user is a space administrator.