SOC-6227 : fix test ExoSpacesAdministrationManageSpaces.test.js (#480)
A test of ExoSpacesAdministrationManageSpaces.test.js fails but it does not make the build fail.
It fails because we did not manage the case where the arguments of the method getSpaceLinkSetting are undefined.
The build did not fail because the test assertions were done in a Promise (nextTick).

This fix handles the case where the arguments of the method getSpaceLinkSetting are undefined by doing a test before using them, all it removes the use of nextTick in the test.
The call to nextTick is not necessary here since we don't do any async task, we just do DOM updates (by updating the variable 'spaces') and vue-test-utils updates DOM synchronously. So the use of nextTick has been removed.
And for information, in the case a call to nextTick is necessary (for example when an async task is required, like fetching data), Jest cannot know when the test is done since it is processed asynchronously, so it must be warn when the test is really done. This must be done by using the 'done()' method, for example:

  it('should display 2 spaces in table when 2 spaces in data', (done) => {
    cmp.vm.spaces = data.spaces;
    cmp.vm.$nextTick(() => {
      const spacesTableRows = cmp.findAll('.manageSpaces table tr');
      expect(spacesTableRows).toHaveLength(3); // header + 2 rows
      done();
    });
  });

See
https://vue-test-utils.vuejs.org/guides/testing-async-components.html for more details.