PLF-6637 : do not use SpacesAdministrationService component since it does not exist in 5.1
The fix for PLF-6637 uses the SpacesAdministrationService component to check if the user is a space administrator or not.
This component has been added in 5.2 and is therefore not available in 5.1.

This fix removes the use of the component SpacesAdministrationService, so the new rule for group selector in space creation is: can only see the groups of the spaces where he is member.