PLF-8419: Fix Empty Space Email Notification
This change is needed because when someone is invited in a space, the mail notification is empty (and there is an error in platform log)
Prior to this change, the notification template use a variable named USER. But this variable is not present in the message context.
To address this issue, I replace the variable USER by SPACE : in fact, the variable USER was used to display a tooltip on the space avatar.