SOC-6175 : make sure that the Loading Icon is not displayed when clicking on Load More button (#567) (#491)
I noticed that there are some activities with no updated date (updatedDate=0 in the table SOC_ACTIVITIES and SOC_STREAM_ITEMS) that are returned at the end of activities and which are comments, so when accessing the activity stream those activities are returned at the end of the request's result (ORDER By DESC) and are confused with activities and aren't loaded completely.
I made sure that comments get an updatedDate and to not retrieve those comments as activities.
After fixing the SQL request to retrieve only streamItems with Stream.type = (Poster or Space), i found out that each post on a space has two entries in SOC_STREAM_ITEMS table: one with the space ID as owner_id and the second one with the poster ID as owner_id.
When a comment is added on that post, the UPDATED_DATE is updated only for the first entry which is with the space ID as owner_id. And so coming to the process of retrieving activity ids (SELECT distinct ACTIVITY_ID, UPDATED_DATE ... ) the two entries are returned, and coming to the process of filtering activity ids the list is going to be reduced because of duplicated ids and so the load more button won't be displayed.
So as a solution i made sure that the UPDATED_DATE is updated for both entries when adding a comment on a space activity.