SOC-6249: Make sure to execute mention notification for newly added mentions only (#552)
Make sure to execute mention notification only when a new mention is added or replaced.
For so at the level of the ActivityManager, i made sure to retrieve the previousMentions of the concerned activity/comment just before updating it on DB and then store them in the template params of the activity to be passed to the Notificaion service.
At the level of the ActivityMentionPlugin i extarct the previous mentions from the activity's template params and using previousMentions and actualMentions variables i deduct the addedMentions by the method getAddedMentions. And so only the addedMentions will be used by the method sendToMentioners().