SOC-6232 : Wrong notification content when react to activity after mention user. (#521)
Prior to this change, a notification can be sent, saying "An activity was updated" when a user like an activity.
This is due to the fact that ActivityManager.updateActivity is called in saveLike and deleteLike. So, updateActivity listeners are launched
This change add a broadcast parameter in updateActivity, to be able to update an activity without broadcasting the event.