SOC-6150 : BindingRules_US05: Leave space
When I'm bound user, I shouldn't be able to leave the space i.e. the leave button is disabled.
This change add a function in Utils class to know if the current user is bound to the space in parameter
In template displaying spaces list (AllSpaces, MySpaces), if user is bound to the space, we disable the leave button, and remove the onlick action on it