PLF-8746 : Perf problem when reading commentsId from an activity with a log of comments
When getting commentId for an activity, we make 2 queries
- 1 to get comments from first level (activities where parentId = activityId)
- 1 to get subcomment (activities where parentId in (commentsId))

If an activity have a lot of comments (like space activity : all comments are member join)
the commentIds list can be huge.

When we make large user add in space, this particular request (with IN) is cached in SessionFactoryImpl.cacheQueryPlan. Each request with different list of id is one entry in cache.
In some case, it can fill the JVM memory.

We replace theses 2 requests by only one : we search activity where parentId=activityId of where parent.parentId=activityId
We search comments and subComments in one request, which can be cached once in cacheQueryPlan.